refactor: split nn into diff files
