+++ python/stick/nn/basic.py
@@ -1,0 +1,300 @@
+from .. import init, ops
+from ..ops import Tensor
+import numpy as np
+from typing import List
+
+class Parameter(Tensor):
+    """A special kind of tensor that represents parameters."""
+
+
+# trace all the fields of the Parameter type
+def _unpack_params(value: object) -> List[Tensor]:
+    if isinstance(value, Parameter):
+        return [value]
+    elif isinstance(value, Module):
+        return value.parameters()
+    elif isinstance(value, dict):
+        params = []
+        for k, v in value.items():
+            params += _unpack_params(v)
+        return params
+    elif isinstance(value, (list, tuple)):
+        params = []
+        for v in value:
+            params += _unpack_params(v)
+        return params
+    else:
+        return []
+
+
+def _child_modules(value: object) -> List["Module"]:
+    if isinstance(value, Module):
+        modules = [value]
+        modules.extend(_child_modules(value.__dict__))
+        return modules
+    if isinstance(value, dict):
+        modules = []
+        for k, v in value.items():
+            modules += _child_modules(v)
+        return modules
+    elif isinstance(value, (list, tuple)):
+        modules = []
+        for v in value:
+            modules += _child_modules(v)
+        return modules
+    else:
+        return []
+
+
+class Module:
+    def __init__(self):
+        self.training = True
+
+    def parameters(self) -> List[Tensor]:
+        """Return the list of parameters in the module."""
+        return _unpack_params(self.__dict__)
+
+    def _children(self) -> List["Module"]:
+        return _child_modules(self.__dict__)
+
+    def eval(self):
+        self.training = False
+        for m in self._children():
+            m.training = False
+
+    def train(self):
+        self.training = True
+        for m in self._children():
+            m.training = True
+
+    def __call__(self, *args, **kwargs):
+        return self.forward(*args, **kwargs)
+
+
+class Identity(Module):
+    def forward(self, x):
+        return x
+
+
+class Linear(Module):
+    def __init__(
+        self, in_features, out_features, bias=True, device=None, dtype="float32"
+    ):
+        super().__init__()
+        self.in_features = in_features
+        self.out_features = out_features
+
+        ### BEGIN YOUR SOLUTION
+        self.weight = Parameter(init.kaiming_uniform(in_features, out_features, requires_grad=True, device=device))
+        self.bias = Parameter(init.kaiming_uniform(out_features, 1, requires_grad=True, device=device).transpose()) if bias else None
+        ### END YOUR SOLUTION
+
+    def forward(self, X: Tensor) -> Tensor:
+        ### BEGIN YOUR SOLUTION
+        out = ops.matmul(X, self.weight)
+        if self.bias:
+            out += ops.broadcast_to(self.bias, out.shape)
+        return out
+        ### END YOUR SOLUTION
+
+
+class Flatten(Module):
+    def forward(self, X):
+        ### BEGIN YOUR SOLUTION
+        len = 1
+        for shape in X.shape[1:]:
+            len *= shape
+        return X.reshape((X.shape[0], len))
+        ### END YOUR SOLUTION
+
+
+class ReLU(Module):
+    def forward(self, x: Tensor) -> Tensor:
+        ### BEGIN YOUR SOLUTION
+        return ops.relu(x)
+        ### END YOUR SOLUTION
+
+
+class Tanh(Module):
+    def forward(self, x: Tensor) -> Tensor:
+        ### BEGIN YOUR SOLUTION
+        return ops.tanh(x)
+        ### END YOUR SOLUTION
+
+
+class Sigmoid(Module):
+    def __init__(self):
+        super().__init__()
+
+    def forward(self, x: Tensor) -> Tensor:
+        ### BEGIN YOUR SOLUTION
+        return (1 + ops.exp(-x)) ** (-1)
+        ### END YOUR SOLUTION
+
+
+class Sequential(Module):
+    def __init__(self, *modules):
+        super().__init__()
+        self.modules = modules
+
+    def forward(self, x: Tensor) -> Tensor:
+        ### BEGIN YOUR SOLUTION
+        for module in self.modules:
+            x = module.forward(x)
+        return x
+        ### END YOUR SOLUTION
+
+
+class SoftmaxLoss(Module):
+    def forward(self, logits: Tensor, y: Tensor):
+        ### BEGIN YOUR SOLUTION
+        return (ops.logsumexp(logits, (1,)) / logits.shape[0]).sum() - (logits * init.one_hot(logits.shape[1], y, device=logits.device) / logits.shape[0]).sum()
+        ### END YOUR SOLUTION
+
+
+class BatchNorm1d(Module):
+    def __init__(self, dim, eps=1e-5, momentum=0.1, device=None, dtype="float32"):
+        super().__init__()
+        self.dim = dim
+        self.eps = eps
+        self.momentum = momentum
+        ### BEGIN YOUR SOLUTION
+        self.weight = Parameter(init.ones(dim, requires_grad=True, device=device))
+        self.bias = Parameter(init.zeros(dim, requires_grad=True, device=device))
+        self.running_mean = init.zeros(dim, device=device) 
+        self.running_var = init.ones(dim, device=device)
+        ### END YOUR SOLUTION
+
+    def forward(self, x: Tensor) -> Tensor:
+        ### BEGIN YOUR SOLUTION
+        if self.training:
+            Ex = (x.sum(axes=(0,)) / x.shape[0])
+            Ex_vec = Ex.reshape((1, x.shape[1]))
+            Varx = (((x + (-Ex_vec).broadcast_to(x.shape)) ** 2).sum(axes=(0,)) / x.shape[0])
+            Varx_vec= Varx.reshape((1, x.shape[1]))
+            self.running_mean = (1 - self.momentum) * self.running_mean + self.momentum * Ex.data
+            self.running_var = (1 - self.momentum) * self.running_var + self.momentum * Varx.data
+            norm = (x + (-Ex_vec).broadcast_to(x.shape)) / ((Varx_vec.broadcast_to(x.shape) + self.eps) ** 0.5)
+            return self.weight.reshape((1, x.shape[1])).broadcast_to(x.shape) * norm + self.bias.reshape((1, x.shape[1])).broadcast_to(x.shape)
+        else:
+            norm = (x + (-self.running_mean).broadcast_to(x.shape)) / ((self.running_var.broadcast_to(x.shape) + self.eps) ** 0.5)
+            return self.weight.reshape((1, x.shape[1])).broadcast_to(x.shape) * norm + self.bias.reshape((1, x.shape[1])).broadcast_to(x.shape)
+        ### END YOUR SOLUTION
+
+
+class BatchNorm2d(BatchNorm1d):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def forward(self, x: Tensor):
+        # nchw -> nhcw -> nhwc
+        s = x.shape
+        _x = x.transpose((1, 2)).transpose((2, 3)).reshape((s[0] * s[2] * s[3], s[1]))
+        y = super().forward(_x).reshape((s[0], s[2], s[3], s[1]))
+        return y.transpose((2,3)).transpose((1,2))
+
+
+class LayerNorm1d(Module):
+    def __init__(self, dim, eps=1e-5, device=None, dtype="float32"):
+        super().__init__()
+        self.dim = dim
+        self.eps = eps
+        ### BEGIN YOUR SOLUTION
+        self.weight = Parameter(init.ones(dim, requires_grad=True, device=device))
+        self.bias = Parameter(init.zeros(dim, requires_grad=True, device=device))
+        ### END YOUR SOLUTION
+
+    def forward(self, x: Tensor) -> Tensor:
+        ### BEGIN YOUR SOLUTION
+        Ex = (x.sum(axes=(1,)) / x.shape[1]).reshape((x.shape[0], 1)).broadcast_to(x.shape)
+        Varx = (((x + (-Ex)) ** 2).sum(axes=(1,)) / x.shape[1]).reshape((x.shape[0], 1)).broadcast_to(x.shape)
+        norm = (x + (-Ex)) / ((Varx + self.eps) ** 0.5)
+        return self.weight.broadcast_to(x.shape) * norm + self.bias.broadcast_to(x.shape)
+        ### END YOUR SOLUTION
+
+
+class Dropout(Module):
+    def __init__(self, p=0.5):
+        super().__init__()
+        self.p = p
+
+    def forward(self, x: Tensor) -> Tensor:
+        ### BEGIN YOUR SOLUTION
+        if not self.training:
+            return x
+        randrop = init.randb(x.shape[0], x.shape[1], p=1 - self.p, device=x.device, dtype=x.dtype)
+        return x * randrop / (1 - self.p)
+        ### END YOUR SOLUTION
+
+
+class Residual(Module):
+    def __init__(self, fn: Module):
+        super().__init__()
+        self.fn = fn
+
+    def forward(self, x: Tensor) -> Tensor:
+        ### BEGIN YOUR SOLUTION
+        return self.fn(x) + x
+        ### END YOUR SOLUTION
+
+class RNNCell(Module):
+    def __init__(self, input_size, hidden_size, bias=True, nonlinearity='tanh', device=None, dtype="float32"):
+        """
+        Applies an RNN cell with tanh or ReLU nonlinearity.
+
+        Parameters:
+        input_size: The number of expected features in the input X
+        hidden_size: The number of features in the hidden state h
+        bias: If False, then the layer does not use bias weights
+        nonlinearity: The non-linearity to use. Can be either 'tanh' or 'relu'.
+
+        Variables:
+        W_ih: The learnable input-hidden weights of shape (input_size, hidden_size).
+        W_hh: The learnable hidden-hidden weights of shape (hidden_size, hidden_size).
+        bias_ih: The learnable input-hidden bias of shape (hidden_size,).
+        bias_hh: The learnable hidden-hidden bias of shape (hidden_size,).
+
+        Weights and biases are initialized from U(-sqrt(k), sqrt(k)) where k = 1/hidden_size
+        """
+        super().__init__()
+        ### BEGIN YOUR SOLUTION
+        bound = np.sqrt(1.0 / hidden_size)
+        self.W_ih = Parameter(init.rand(input_size, hidden_size, low=-bound, high=bound, device=device, dtype=dtype, requires_grad=True))
+        self.W_hh = Parameter(init.rand(hidden_size, hidden_size, low=-bound, high=bound, device=device, dtype=dtype, requires_grad=True))
+        if bias:
+            self.bias_ih = Parameter(init.rand(hidden_size, low=-bound, high=bound, device=device, dtype=dtype, requires_grad=True))
+            self.bias_hh = Parameter(init.rand(hidden_size, low=-bound, high=bound, device=device, dtype=dtype, requires_grad=True))
+        else:
+            self.bias_ih = None
+            self.bias_hh = None
+        if nonlinearity == 'tanh':
+            self.nonlinearity = Tanh()
+        elif nonlinearity == 'relu':
+            self.nonlinearity = ReLU()
+        self.bias = bias
+        self.device = device
+        self.dtype = dtype
+        self.hidden_size = hidden_size
+        ### END YOUR SOLUTION
+
+    def forward(self, X, h=None):
+        """
+        Inputs:
+        X of shape (bs, input_size): Tensor containing input features
+        h of shape (bs, hidden_size): Tensor containing the initial hidden state
+            for each element in the batch. Defaults to zero if not provided.
+
+        Outputs:
+        h' of shape (bs, hidden_size): Tensor contianing the next hidden state
+            for each element in the batch.
+        """
+        ### BEGIN YOUR SOLUTION
+        bs = X.shape[0]
+        if h is None:
+            h = init.zeros(bs, self.hidden_size, device=self.device, dtype=self.dtype)
+        if self.bias:
+            return self.nonlinearity(X @ self.W_ih + self.bias_ih.reshape((1, self.hidden_size)).broadcast_to((bs, self.hidden_size)) + 
+                                     h @ self.W_hh + self.bias_hh.reshape((1, self.hidden_size)).broadcast_to((bs, self.hidden_size)))
+        else:
+            return self.nonlinearity(X @ self.W_ih + h @ self.W_hh)
+        ### END YOUR SOLUTION
