+++ python/stick/nn/rnn.py
@@ -1,0 +1,246 @@
+from .basic import *
+
+
+class RNN(Module):
+    def __init__(self, input_size, hidden_size, num_layers=1, bias=True, nonlinearity='tanh', device=None, dtype="float32"):
+        """
+        Applies a multi-layer RNN with tanh or ReLU non-linearity to an input sequence.
+
+        Parameters:
+        input_size - The number of expected features in the input x
+        hidden_size - The number of features in the hidden state h
+        num_layers - Number of recurrent layers.
+        nonlinearity - The non-linearity to use. Can be either 'tanh' or 'relu'.
+        bias - If False, then the layer does not use bias weights.
+
+        Variables:
+        rnn_cells[k].W_ih: The learnable input-hidden weights of the k-th layer,
+            of shape (input_size, hidden_size) for k=0. Otherwise the shape is
+            (hidden_size, hidden_size).
+        rnn_cells[k].W_hh: The learnable hidden-hidden weights of the k-th layer,
+            of shape (hidden_size, hidden_size).
+        rnn_cells[k].bias_ih: The learnable input-hidden bias of the k-th layer,
+            of shape (hidden_size,).
+        rnn_cells[k].bias_hh: The learnable hidden-hidden bias of the k-th layer,
+            of shape (hidden_size,).
+        """
+        super().__init__()
+        ### BEGIN YOUR SOLUTION
+        self.rnn_cells = [RNNCell(input_size, hidden_size, bias=bias, nonlinearity=nonlinearity, device=device, dtype=dtype)] + \
+                         [RNNCell(hidden_size, hidden_size, bias=bias, nonlinearity=nonlinearity, device=device, dtype=dtype) for _ in range(num_layers - 1)]
+        self.num_layers = num_layers
+        self.hidden_size = hidden_size
+        self.device = device
+        self.dtype = dtype
+        ### END YOUR SOLUTION
+
+    def forward(self, X, h0=None):
+        """
+        Inputs:
+        X of shape (seq_len, bs, input_size) containing the features of the input sequence.
+        h_0 of shape (num_layers, bs, hidden_size) containing the initial
+            hidden state for each element in the batch. Defaults to zeros if not provided.
+
+        Outputs
+        output of shape (seq_len, bs, hidden_size) containing the output features
+            (h_t) from the last layer of the RNN, for each t.
+        h_n of shape (num_layers, bs, hidden_size) containing the final hidden state for each element in the batch.
+        """
+        ### BEGIN YOUR SOLUTION
+        X_split = list(tuple(ops.split(X, axis=0)))
+        if h0 is None:
+            h0 = init.zeros(self.num_layers, X.shape[1], self.hidden_size, device=self.device, dtype=self.dtype)
+        h_n = list(tuple(ops.split(h0, axis=0)))
+        for num_layer in range(self.num_layers):
+            h_t = h_n[num_layer]
+            for i in range(X.shape[0]):
+                X_split[i] = self.rnn_cells[num_layer](X_split[i], h_t)
+                h_t = X_split[i]
+            h_n[num_layer] = h_t
+        return ops.stack(X_split, axis=0), ops.stack(h_n, axis=0)
+        ### END YOUR SOLUTION
+
+
+
+class LSTMCell(Module):
+    def __init__(self, input_size, hidden_size, bias=True, device=None, dtype="float32"):
+        """
+        A long short-term memory (LSTM) cell.
+
+        Parameters:
+        input_size - The number of expected features in the input X
+        hidden_size - The number of features in the hidden state h
+        bias - If False, then the layer does not use bias weights
+
+        Variables:
+        W_ih - The learnable input-hidden weights, of shape (input_size, 4*hidden_size).
+        W_hh - The learnable hidden-hidden weights, of shape (hidden_size, 4*hidden_size).
+        bias_ih - The learnable input-hidden bias, of shape (4*hidden_size,).
+        bias_hh - The learnable hidden-hidden bias, of shape (4*hidden_size,).
+
+        Weights and biases are initialized from U(-sqrt(k), sqrt(k)) where k = 1/hidden_size
+        """
+        super().__init__()
+        ### BEGIN YOUR SOLUTION
+        bound = np.sqrt(1.0 / hidden_size)
+        self.W_ih = Parameter(init.rand(input_size, 4 * hidden_size, low=-bound, high=bound, device=device, dtype=dtype, requires_grad=True))
+        self.W_hh = Parameter(init.rand(hidden_size, 4 * hidden_size, low=-bound, high=bound, device=device, dtype=dtype, requires_grad=True))
+        if bias:
+            self.bias_ih = Parameter(init.rand(4 * hidden_size, low=-bound, high=bound, device=device, dtype=dtype, requires_grad=True))
+            self.bias_hh = Parameter(init.rand(4 * hidden_size, low=-bound, high=bound, device=device, dtype=dtype, requires_grad=True))
+        else:
+            self.bias_ih = None
+            self.bias_hh = None
+        self.bias = bias
+        self.device = device
+        self.dtype = dtype
+        self.hidden_size = hidden_size
+        ### END YOUR SOLUTION
+
+
+    def forward(self, X, h=None):
+        """
+        Inputs: X, h
+        X of shape (batch, input_size): Tensor containing input features
+        h, tuple of (h0, c0), with
+            h0 of shape (bs, hidden_size): Tensor containing the initial hidden state
+                for each element in the batch. Defaults to zero if not provided.
+            c0 of shape (bs, hidden_size): Tensor containing the initial cell state
+                for each element in the batch. Defaults to zero if not provided.
+
+        Outputs: (h', c')
+        h' of shape (bs, hidden_size): Tensor containing the next hidden state for each
+            element in the batch.
+        c' of shape (bs, hidden_size): Tensor containing the next cell state for each
+            element in the batch.
+        """
+        ### BEGIN YOUR SOLUTION
+        bs = X.shape[0]
+        if h is None:
+            h0 = init.zeros(bs, self.hidden_size, device=self.device, dtype=self.dtype)
+            c0 = init.zeros(bs, self.hidden_size, device=self.device, dtype=self.dtype)
+        else:
+            h0 = h[0]
+            c0 = h[1]
+        if self.bias:
+            gate_input = (X @ self.W_ih + self.bias_ih.reshape((1, 4 * self.hidden_size)).broadcast_to((bs, 4 * self.hidden_size)) +
+                          h0 @ self.W_hh + self.bias_hh.reshape((1, 4 * self.hidden_size)).broadcast_to((bs, 4 * self.hidden_size)))
+        else:
+            gate_input = (X @ self.W_ih + h0 @ self.W_hh)
+        gate_input_split = tuple(ops.split(gate_input, axis=1))
+        i, f, g, o = Sigmoid()(ops.stack(gate_input_split[0 : self.hidden_size], axis=1)), \
+                    Sigmoid()(ops.stack(gate_input_split[self.hidden_size : 2 * self.hidden_size], axis=1)), \
+                    Tanh()(ops.stack(gate_input_split[2 * self.hidden_size:3 * self.hidden_size], axis=1)), \
+                    Sigmoid()(ops.stack(gate_input_split[3 * self.hidden_size:4 * self.hidden_size], axis=1))
+        c_ = f * c0 + i * g
+        h_ = o * Tanh()(c_)
+        return h_, c_
+        ### END YOUR SOLUTION
+
+
+class LSTM(Module):
+    def __init__(self, input_size, hidden_size, num_layers=1, bias=True, device=None, dtype="float32"):
+        super().__init__()
+        """
+        Applies a multi-layer long short-term memory (LSTM) RNN to an input sequence.
+
+        Parameters:
+        input_size - The number of expected features in the input x
+        hidden_size - The number of features in the hidden state h
+        num_layers - Number of recurrent layers.
+        bias - If False, then the layer does not use bias weights.
+
+        Variables:
+        lstm_cells[k].W_ih: The learnable input-hidden weights of the k-th layer,
+            of shape (input_size, 4*hidden_size) for k=0. Otherwise the shape is
+            (hidden_size, 4*hidden_size).
+        lstm_cells[k].W_hh: The learnable hidden-hidden weights of the k-th layer,
+            of shape (hidden_size, 4*hidden_size).
+        lstm_cells[k].bias_ih: The learnable input-hidden bias of the k-th layer,
+            of shape (4*hidden_size,).
+        lstm_cells[k].bias_hh: The learnable hidden-hidden bias of the k-th layer,
+            of shape (4*hidden_size,).
+        """
+        ### BEGIN YOUR SOLUTION
+        self.lstm_cells = [LSTMCell(input_size, hidden_size, bias=bias, device=device, dtype=dtype)] + \
+                         [LSTMCell(hidden_size, hidden_size, bias=bias, device=device, dtype=dtype) for _ in range(num_layers - 1)]
+        self.num_layers = num_layers
+        self.hidden_size = hidden_size
+        self.device = device
+        self.dtype = dtype
+        ### END YOUR SOLUTION
+
+    def forward(self, X, h=None):
+        """
+        Inputs: X, h
+        X of shape (seq_len, bs, input_size) containing the features of the input sequence.
+        h, tuple of (h0, c0) with
+            h_0 of shape (num_layers, bs, hidden_size) containing the initial
+                hidden state for each element in the batch. Defaults to zeros if not provided.
+            c0 of shape (num_layers, bs, hidden_size) containing the initial
+                hidden cell state for each element in the batch. Defaults to zeros if not provided.
+
+        Outputs: (output, (h_n, c_n))
+        output of shape (seq_len, bs, hidden_size) containing the output features
+            (h_t) from the last layer of the LSTM, for each t.
+        tuple of (h_n, c_n) with
+            h_n of shape (num_layers, bs, hidden_size) containing the final hidden state for each element in the batch.
+            h_n of shape (num_layers, bs, hidden_size) containing the final hidden cell state for each element in the batch.
+        """
+        ### BEGIN YOUR SOLUTION
+        X_split = list(tuple(ops.split(X, axis=0)))
+        if h is None:
+            h0 = init.zeros(self.num_layers, X.shape[1], self.hidden_size, device=self.device, dtype=self.dtype)
+            c0 = init.zeros(self.num_layers, X.shape[1], self.hidden_size, device=self.device, dtype=self.dtype)
+        else:
+            h0 = h[0]
+            c0 = h[1]
+        h_n = list(tuple(ops.split(h0, axis=0)))
+        c_n = list(tuple(ops.split(c0, axis=0)))
+        for num_layer in range(self.num_layers):
+            h_t = h_n[num_layer]
+            c_t = c_n[num_layer]
+            for i in range(X.shape[0]):
+                X_split[i], c_t = self.lstm_cells[num_layer](X_split[i], (h_t, c_t))
+                h_t = X_split[i]
+            h_n[num_layer] = h_t
+            c_n[num_layer] = c_t
+        return ops.stack(X_split, axis=0), (ops.stack(h_n, axis=0), ops.stack(c_n, axis=0))
+        ### END YOUR SOLUTION
+
+
+class Embedding(Module):
+    def __init__(self, num_embeddings, embedding_dim, device=None, dtype="float32"):
+        super().__init__()
+        """
+        Maps one-hot word vectors from a dictionary of fixed size to embeddings.
+
+        Parameters:
+        num_embeddings (int) - Size of the dictionary
+        embedding_dim (int) - The size of each embedding vector
+
+        Variables:
+        weight - The learnable weights of shape (num_embeddings, embedding_dim)
+            initialized from N(0, 1).
+        """
+        ### BEGIN YOUR SOLUTION
+        self.weight = Parameter(init.randn(num_embeddings, embedding_dim, mean=0.0, std=1.0, device=device, dtype=dtype, requires_grad=True))
+        self.device = device
+        self.dtype = dtype
+        ### END YOUR SOLUTION
+
+    def forward(self, x: Tensor) -> Tensor:
+        """
+        Maps word indices to one-hot vectors, and projects to embedding vectors
+
+        Input:
+        x of shape (seq_len, bs)
+
+        Output:
+        output of shape (seq_len, bs, embedding_dim)
+        """
+        ### BEGIN YOUR SOLUTION
+        seq_len, bs = x.shape
+        x_one_hot = init.one_hot(self.weight.shape[0], x, device=self.device, dtype=self.dtype).reshape((seq_len * bs, self.weight.shape[0]))
+        return (x_one_hot @ self.weight).reshape((seq_len, bs, self.weight.shape[1]))
+        ### END YOUR SOLUTION
